+++ src/main/groovy/com/exictos/devops/Application.groovy
@@ -1,0 +1,9 @@
+package com.exictos.devops
+
+import com.exictos.devops.helpers.XHDLogger
+
+class Application {
+
+    XHDLogger logger = new XHDLogger()
+
+}