+++ src/main/groovy/com/exictos/devops/containers/WebSphere.groovy
@@ -17,8 +17,8 @@ class WebSphere extends Container{
 
     WebSphere(String host, int port, String username, char[] password, String aWsadmin)
     {
-        wsadmin = new WSAdminWrapper(aWsadmin,host,port,username,password)
-        profile = new WebSphereProfile(wsadmin)
+        wsadmin = new WSAdminWrapper(aWsadmin,host,port,username,password, logger)
+        profile = new WebSphereProfile(wsadmin, logger)
         profile.host = host
         profile.port = port
         profile.username = username
@@ -54,15 +54,16 @@ class WebSphere extends Container{
     protected String installApp(File pathToPackage, String applicationName, String applicationVersion
                                 , Timestamp timestamp)
     {
-        log.info("Installing application ${applicationName} from package at ${pathToPackage}...")
+        logger.log("Installing application ${applicationName} from package at ${pathToPackage}...")
         try{
             String name = LiberoHelper.standardizeName(pathToPackage.getAbsolutePath(), applicationName, applicationVersion
                     , timestamp)
             wsadmin.installApplication(pathToPackage.getAbsolutePath(),name)
-            log.info("${applicationName} installed successfully as ${name}")
+            wsadmin.disableAutoStart(name)
+            logger.log("${applicationName} installed successfully as ${name}")
             return name
         }catch(Exception e){
-            log.error "Could not install application ${applicationName} from package ${pathToPackage}. Cause: ${e.getMessage()}"
+            logger.log "Could not install application ${applicationName} from package ${pathToPackage}. Cause: ${e.getMessage()}"
             throw e
         }
     }
@@ -75,13 +76,13 @@ class WebSphere extends Container{
     @Override
     void startApp(String deploymentName)
     {
-        log.info("Starting application: ${deploymentName}...")
+        logger.log("Starting application: ${deploymentName}...")
         try{
             if(wsadmin.isAppReady(deploymentName))
                 wsadmin.startApplication(deploymentName)
-            log.info("Deployment ${deploymentName} started.")
+            logger.log("Deployment ${deploymentName} started.")
         }catch(Exception e){
-            log.error("Could not start application ${deploymentName}. Cause: ${e}")
+            logger.log("Could not start application ${deploymentName}. Cause: ${e}")
             throw e
         }
     }
@@ -95,12 +96,12 @@ class WebSphere extends Container{
     @Override
     void stopApp(String deploymentName)
     {
-        log.info("Stopping deployment ${deploymentName}...")
+        logger.log("Stopping deployment ${deploymentName}...")
         try{
             wsadmin.stopApplication(deploymentName)
-            log.info("Deployment ${deploymentName} stopped")
+            logger.log("Deployment ${deploymentName} stopped")
         }catch(Exception e){
-            log.error("Could not stop deployment: ${deploymentName}. Cause: ${e}")
+            logger.log("Could not stop deployment: ${deploymentName}. Cause: ${e}")
             throw e
         }
     }
@@ -113,12 +114,44 @@ class WebSphere extends Container{
     @Override
     void uninstallApp(String deploymentName)
     {
-        log.info("Uninstalling deployment ${deploymentName}...")
+        logger.log("Uninstalling deployment ${deploymentName}...")
         try{
             wsadmin.uninstallApp(deploymentName)
-            log.info("Deployment ${deploymentName} uninstalled.")
+            logger.log("Deployment ${deploymentName} uninstalled.")
         }catch(Exception e){
-            log.error "Could not uninstall ${deploymentName}. Cause: ${e}"
+            logger.log "Could not uninstall ${deploymentName}. Cause: ${e}"
+            throw e
+        }
+    }
+
+    /**
+     * Disables application loading when WebSphere boots up
+     *
+     * @param deploymentName
+     */
+    @Override
+    protected void disableAutoStart(String deploymentName) {
+        logger.log("Disabling ${deploymentName} autostart")
+        try{
+            wsadmin.disableAutoStart(deploymentName)
+        }catch(Exception e){
+            logger.log("Could not disable autostart. Cause ${e}")
+            throw e
+        }
+    }
+
+    /**
+     * Enables application loading when WebSphere boots up
+     *
+     * @param deploymentName
+     */
+    @Override
+    protected void enableAutoStart(String deploymentName) {
+        logger.log("Enabling ${deploymentName} autostart")
+        try{
+            wsadmin.enableAutoStart(deploymentName)
+        }catch(Exception e){
+            logger.log("Could not enable autostart. Cause ${e}")
             throw e
         }
     }