+++ src/main/groovy/com/exictos/devops/helpers/WSAdminWrapper.groovy
@@ -1,13 +1,13 @@
 package com.exictos.devops.helpers
 
-import ch.qos.logback.classic.Logger
+import com.exictos.devops.Application
 
 /**
  * This is a wrapper for WebSphere wsadmin script. It encapsulates the basic logic of calling wsadmin.
  */
-class WSAdminWrapper {
+class WSAdminWrapper{
 
-    protected static final Logger log = LiberoLogger.getLogger()
+    protected XHDLogger log
 
     String home
     String host
@@ -15,12 +15,13 @@ class WSAdminWrapper {
     String username
     char[] password
 
-    WSAdminWrapper(String path, String host, int port, String username, char[] password){
+    WSAdminWrapper(String path, String host, int port, String username, char[] password, XHDLogger logger = new Application().getLogger()){
         this.home = path
         this.host = host
         this.port = port
         this.username = username
         this.password = password
+        this.log = logger
     }
 
     /**
@@ -117,6 +118,37 @@ class WSAdminWrapper {
     }
 
     /**
+     *  Enables an application loading in all deployed targets
+     * @param deploymentName
+     */
+    void enableAutoStart(String deploymentName){
+        setApplicationLoading(true, deploymentName)
+    }
+
+    /**
+     *  Disables an application loading in all deployed targets
+     *
+     * @param deploymentName
+     */
+    void disableAutoStart(String deploymentName){
+        setApplicationLoading(false, deploymentName)
+    }
+
+    /**
+     * Enables/Disables an application loading in deployed targets
+     *
+     * @param value true to enable and false to disable
+     * @param deploymentName The application deployment name
+     */
+    private void setApplicationLoading(boolean value, String deploymentName){
+        run("set deployments [\$AdminConfig getid /Deployment:${deploymentName}/];" +
+                "set deploymentObj1 [\$AdminConfig showAttribute \$deployments deployedObject];" +
+                "set targetMap1 [lindex [\$AdminConfig showAttribute \$deploymentObj1 targetMappings] 0];" +
+                "foreach tm \$targetMap1 {\$AdminConfig modify \$tm {{enable ${value}}}}")
+        saveConfig()
+    }
+
+    /**
      * Saves WAS configuration after a management operation is performed.
      */
     void saveConfig()
@@ -135,11 +167,11 @@ class WSAdminWrapper {
                 "-c \"${command}\""
         CmdRunner.CmdResponse response = CmdRunner.getResponse(cmd)
         if(response.exitCode != 0){
-            log.error(response.getOutput().trim())
+            log.log(response.getOutput().trim())
             throw new RuntimeException("WSAdmin command returned non-zero code(${response.getExitCode()})")
         }
 
-        log.info(response.getOutput().trim())
+        log.log(response.getOutput().trim())
         return response.getOutput().trim()
     }