+++ src/main/groovy/com/exictos/devops/services/managers/ServiceManager.groovy
@@ -1,27 +1,27 @@
 package com.exictos.devops.services.managers
 
-import ch.qos.logback.classic.Logger
+
+import com.exictos.devops.Application
 import com.exictos.devops.helpers.LiberoHelper
-import com.exictos.devops.helpers.LiberoLogger
 import com.exictos.devops.helpers.NssmWrapper
+import com.exictos.devops.helpers.XHDLogger
 import com.exictos.devops.profiles.Instance
 import com.exictos.devops.services.Service
 
 /**
  * The abstract class representing a service manager. Should be extended by concrete classes for each supported OS.
  */
-abstract class ServiceManager {
+abstract class ServiceManager{
 
-    protected static final Logger log = LiberoLogger.getLogger()
-
+    protected XHDLogger logger = new Application().getLogger()
     /**
-     * Sets the file path to log events
+     * Sets the file path to logger events
      *
-     * @param the full filePath (eg. C:/logs/output.log)
+     * @param the full filePath (eg. C:/logs/output.logger)
      */
-    static void setLogFile(File filePath){
-        LiberoLogger.setLogFile(filePath.toString())
-        log.debug("Logging to ${filePath.toString()}")
+    void setLogFile(File filePath){
+        logger.setLogFile(filePath.toString())
+        logger.log("Logging to ${filePath}")
     }
 
     /**
@@ -70,9 +70,9 @@ abstract class ServiceManager {
      * @param service
      * @return a list of instances of service installed
      */
-    static List<Instance> listInstances(Service service)
+    List<Instance> listInstances(Service service)
     {
-        log.info("Listing instances of ${service.getName()}...")
+        logger.log("Listing instances of ${service.getName()}...")
         List<Instance> instances = new ArrayList<Instance>()
         String instancePrefix = LiberoHelper.extractFolderNameFromPackageFile(service._package)
 
@@ -80,7 +80,11 @@ abstract class ServiceManager {
             if(directory.getName().startsWith(instancePrefix) && LiberoHelper.isValidDeploymentName(directory.getName())){
                 Instance instance = new Instance()
                 instance.setName(directory.getName())
-                instance.setTimestamp(LiberoHelper.extractTimestamp(directory.getName()))
+                try{
+                    instance.setTimestamp(LiberoHelper.extractTimestamp(directory.getName()))
+                }catch(Exception e){
+                    logger.log("Could not parse application timestamp. Cause: ${e}")
+                }
                 instances.add(instance)
             }
         }
@@ -95,7 +99,7 @@ abstract class ServiceManager {
      */
     void installServiceWithRollback(Service service)
     {
-        log.info("Installing service ${service.getName()} with rollback...")
+        logger.log("Installing service ${service.getName()} with rollback")
         stop(service)
         uninstallOldInstances(service)
         installService(service)
@@ -114,9 +118,9 @@ abstract class ServiceManager {
     * @param service
     * @param oldnessThreshold
     */
-    static void uninstallOldInstances(Service service, int oldnessThreshold = 0)
+    void uninstallOldInstances(Service service, int oldnessThreshold = 0)
     {
-        log.info("uninstalling old instances of ${service.getName()}...")
+        logger.log("Uninstalling old instances of ${service.getName()}...")
         listInstances(service).each {instance ->
             if(instance.getOldness() > oldnessThreshold)
                 new File(service.installDirectory,instance.getName()).deleteDir()